+++ characters.h
@@ -1,0 +1,10 @@
+#ifndef CHARACTERS_H
+#define CHARACTERS_H
+
+extern int ch;
+extern int player_x, player_y, max_x, max_y;
+extern char *player_01, *player_02;
+
+void player_control();
+
+#endif
