+++ game_loop.h
@@ -1,0 +1,10 @@
+#ifndef GAME_LOOP_H
+#define GAME_LOOP_H
+
+extern int ch;
+extern int player_x, player_y, max_x, max_y;
+extern char *player_01 , *player_02;
+
+void execute_until_esc();
+
+#endif
